requirements.txt
numpy==1.24.3
pandas==1.5.3
pyarrow==12.0.0
streamlit
pickle-mixin
scikit-learn==1.3.2
openai
ipykernel
psycopg2-binary
python-dotenv
tqdm
fastapi
uvicorn
pydantic
apscheduler
sqlalchemy